+++ src/Localization/Localization.mjs
@@ -1,3 +1,3 @@
 /**
- * @typedef {{getLanguage: (module: string, language: string) => Promise<{direction: string}>, getLanguages: (module: string, exclude_system: boolean) => Promise<{[key: string]: string}>, translate: (module: string, key: string, placeholders: null, language: string) => Promise<string>}} Localization
+ * @typedef {{getLanguage: (language: string) => Promise<{direction: string}>, getLanguages: (exclude_system: boolean) => Promise<{[key: string]: string}>, translate: (module: string, key: string, placeholders: null, language: string) => Promise<string>}} Localization
  */